# vendor-user-api/public/swagger.yaml
openapi: 3.0.0
info:
  title: Vendor User API
  description: |
    REST API for managing users, vendors, events, messages, reviews, and notifications.
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer YOUR_JWT_TOKEN
    ```
    
    Tokens expire after 1 hour (3600 seconds).
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://127.0.0.1:8000/api
    description: Local development server
  - url: http://localhost:8000/api
    description: Alternative local server

tags:
  - name: Authentication
    description: User authentication and token management
  - name: Users
    description: User management operations
  - name: Vendors
    description: Vendor profile management
  - name: Events
    description: Event booking management
  - name: Messages
    description: Messaging between users
  - name: Reviews
    description: Vendor review management
  - name: Notifications
    description: User notifications

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        fullName:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "1234567890"
        address:
          type: string
          example: "123 Main St"
        role:
          type: string
          enum: [Customer, Vendor, Admin]
          example: "Customer"
        area:
          type: string
          example: "Downtown"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Vendor:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        businessName:
          type: string
        contactName:
          type: string
        businessPhone:
          type: string
        businessEmail:
          type: string
        businessAddress:
          type: string
        serviceArea:
          type: string
        serviceType:
          type: string
        cuisineStyle:
          type: string
        description:
          type: string
        website:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        linkedin:
          type: string
        ratingAverage:
          type: number
          format: double
        reviewCount:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Event:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        vendor_id:
          type: integer
        eventType:
          type: string
        eventDate:
          type: string
          format: date
        description:
          type: string
        dietaryRestrictions:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object

paths:
  /Users/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - email
                - password
              properties:
                fullName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                phone:
                  type: string
                address:
                  type: string
                role:
                  type: string
                  enum: [Customer, Vendor]
                area:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'

  /Users/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me:
    get:
      tags:
        - Authentication
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /Users:
    get:
      tags:
        - Users
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

  /Users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    
    put:
      tags:
        - Users
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                phone:
                  type: string
                address:
                  type: string
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /Vendors/searchall:
    get:
      tags:
        - Vendors
      summary: Search vendors
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: searchText
          in: query
          schema:
            type: string
        - name: isOffer
          in: query
          schema:
            type: boolean
        - name: isFavorite
          in: query
          schema:
            type: boolean
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vendor'

  /Vendors/{id}:
    get:
      tags:
        - Vendors
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Vendor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
    
    put:
      tags:
        - Vendors
      summary: Update vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                businessName:
                  type: string
                description:
                  type: string
                serviceArea:
                  type: string
      responses:
        '200':
          description: Vendor updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  vendor:
                    $ref: '#/components/schemas/Vendor'

  /events:
    get:
      tags:
        - Events
      summary: Get all events
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    
    post:
      tags:
        - Events
      summary: Create event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - eventType
                - eventDate
              properties:
                user_id:
                  type: integer
                vendor_id:
                  type: integer
                eventType:
                  type: string
                eventDate:
                  type: string
                  format: date
                description:
                  type: string
                dietaryRestrictions:
                  type: string
                budgetMin:
                  type: number
                budgetMax:
                  type: number
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  event:
                    $ref: '#/components/schemas/Event'

  /events/{id}:
    get:
      tags:
        - Events
      summary: Get event by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
    
    put:
      tags:
        - Events
      summary: Update event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                eventDate:
                  type: string
                  format: date
                description:
                  type: string
                budgetMax:
                  type: number
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  event:
                    $ref: '#/components/schemas/Event'
    
    delete:
      tags:
        - Events
      summary: Delete event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /messages/send:
    post:
      tags:
        - Messages
      summary: Send message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - senderId
                - receiverId
                - message
              properties:
                senderId:
                  type: integer
                receiverId:
                  type: integer
                message:
                  type: string
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

  /reviews:
    post:
      tags:
        - Reviews
      summary: Save review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
                - rating
              properties:
                eventId:
                  type: integer
                rating:
                  type: number
                title:
                  type: string
                comment:
                  type: string
      responses:
        '200':
          description: Review saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string